requirements.txt
Flask==3.0.0
python-docx==0.8.11
PyPDF2==3.0.1
Werkzeug==3.0.1
requests==2.31.0
